Feat/Added sounds, sprites and grid behaviours

+++ Minigin/FHTime.cpp
@@ -2,10 +2,10 @@
 
 void FH::Time::UpdateTime(const std::chrono::steady_clock::time_point& time)
 {
-	const std::chrono::duration<double, std::milli> deltaMillis{(time - PREVTIME)};
+	const std::chrono::duration<float, std::milli> deltaMillis{(time - PREVTIME)};
 	PREVTIME = time;
 
-	DELTATIME = deltaMillis.count() / 1000.0;
+	DELTATIME = deltaMillis.count() / 1000.0f;
 
 	ELAPSEDTIME += DELTATIME;
 }
