+++ Minigin/FHTime.h
@@ -6,8 +6,8 @@ namespace FH
 	class Time final
 	{
 	public:
-		static double GetDeltaTime() { return DELTATIME; }
-		static double GetElapsedTime() { return ELAPSEDTIME; }
+		static float GetDeltaTime() { return DELTATIME; }
+		static float GetElapsedTime() { return ELAPSEDTIME; }
 
 		static void UpdateTime(const std::chrono::steady_clock::time_point& time);
 
@@ -19,8 +19,8 @@ namespace FH
 		Time& operator=(Time&& other) = delete;
 
 	private:
-		inline static double DELTATIME{ 0.01 };
-		inline static double ELAPSEDTIME{ 0.01 };
+		inline static float DELTATIME{ 0.01f };
+		inline static float ELAPSEDTIME{ 0.01f };
 
 		inline static std::chrono::steady_clock::time_point 
 			PREVTIME{ std::chrono::steady_clock::now() };
